[ UPDATE ] TaskTestView - sum float support

+++ apps/usuario/views.py
@@ -13,7 +13,7 @@ class TaskTestView(TemplateView):
 
     def get_context_data(self, **kwargs):
         numbers_raw = self.request.GET.get("sum", "1,1")
-        numbers = map(int, numbers_raw.split(","))
+        numbers = map(float, numbers_raw.split(","))
         async_task = test_task.apply_async(
             countdown=5, args=list(numbers)
         )
